Refactor and enhance interval tests for improved coverage and clarity
