+++ python/tests/intervals/core/interval_tests.py
@@ -56,8 +56,8 @@ class TestInterval:
         data = Series({"start": None, "end": "2023-01-02", 1: "A", "metric1": 10, "metric2": 15})
 
         with pytest.raises(InvalidSeriesColumnError, match="All series_fields must be strings"):
-            interval = Interval.create(data, start_field="start", end_field="end", series_fields=[1],
-                                       metric_fields=["metric1", "metric2"])
+            Interval.create(data, start_field="start", end_field="end", series_fields=[1],
+                            metric_fields=["metric1", "metric2"])
 
     def test_create_interval_series_fields_not_sequence(self):
         data = Series({"start": "2023-01-01", "end": "2023-01-05", "series-id1": "ABC"})
@@ -68,25 +68,21 @@ class TestInterval:
         data = Series({"start": None, "end": "2023-01-02", "series": "A", 1: 10, "metric": 15})
 
         with pytest.raises(InvalidMetricColumnError, match="All metric_fields must be strings"):
-            interval = Interval.create(data, start_field="start", end_field="end", series_fields=["series"],
-                                       metric_fields=[1, "metric"])
+            Interval.create(data, start_field="start", end_field="end", series_fields=["series"],
+                            metric_fields=[1, "metric"])
 
     def test_create_interval_metrics_fields_not_sequence(self):
         data = Series({"start": None, "end": "2023-01-02", "series": "A", 1: 10, "metric2": 15})
 
         with pytest.raises(InvalidMetricColumnError, match="metric_fields must be a sequence"):
-            interval = Interval.create(data, start_field="start", end_field="end", series_fields=["series"],
-                                       metric_fields="metric")
+            Interval.create(data, start_field="start", end_field="end", series_fields=["series"],
+                            metric_fields="metric")
 
     def test_invalid_data_type(self):
         data = {"start": "2023-01-01", "end": "2023-01-02"}  # Not a pandas Series
 
-        # Explicitly verify the exception is raised
-        try:
+        with pytest.raises(InvalidDataTypeError, match="Data must be a pandas Series"):
             Interval.create(data, start_field="start", end_field="end")
-            pytest.fail("Expected InvalidDataTypeError was not raised")
-        except InvalidDataTypeError as e:
-            assert str(e) == "Data must be a pandas Series"
 
     def test_empty_data(self):
         data = Series(dtype="object")  # Empty pandas Series
@@ -143,7 +139,6 @@ class TestInterval:
         assert result.is_valid
 
     def test_validate_metric_alignment_invalid(self):
-
         data1 = Series(
             {"start": "2023-01-01", "end": "2023-01-05", "series1": "A", "series2": "B", "metric1": 10, "metric2": 15})
         interval1 = Interval.create(data1, start_field="start", end_field="end", metric_fields=["metric1", "metric2"])
@@ -152,10 +147,7 @@ class TestInterval:
         interval2 = Interval.create(data2, start_field="start", end_field="end", metric_fields=["metric1"])
 
         expected_msg = re.escape("metric_fields don't match: ['metric1', 'metric2'] vs ['metric1']")
-        with pytest.raises(
-                InvalidMetricColumnError,
-                match=expected_msg
-        ):
+        with pytest.raises(InvalidMetricColumnError, match=expected_msg):
             interval1.validate_metrics_alignment(interval2)
 
     def test_validate_series_alignment(self):
@@ -182,19 +174,14 @@ class TestInterval:
                                     metric_fields=["metric1", "metric2"])
 
         expected_msg = re.escape("series_fields don't match: ['series1'] vs ['series1', 'series2']")
-        with pytest.raises(
-                InvalidSeriesColumnError,
-                match=expected_msg):
+        with pytest.raises(InvalidSeriesColumnError, match=expected_msg):
             interval1.validate_series_alignment(interval2)
 
     def test_validate_not_point_in_time_valid_interval(self):
         data = Series({"start_time": 1, "end_time": 2})
         boundary_accessor = _BoundaryAccessor("start_time", "end_time")
 
-        try:
-            Interval._validate_not_point_in_time(data, boundary_accessor)
-        except InvalidDataTypeError:
-            pytest.fail("InvalidDataTypeError raised for a valid interval")
+        Interval._validate_not_point_in_time(data, boundary_accessor)
 
     def test_validate_not_point_in_time_point_in_time_interval(self):
         data = Series({"start_time": 1, "end_time": 1})
@@ -209,3 +196,95 @@ class TestInterval:
 
         with pytest.raises(KeyError):
             Interval._validate_not_point_in_time(data, boundary_accessor)
+
+
+class TestStillValidLegacy:
+
+    def test_update_interval_boundary_start(self):
+        interval = Interval.create(Series(
+            {"start": "2023-01-01T01:00:00", "end": "2023-01-01T02:00:00"}
+        ), "start", "end")
+
+        updated = interval.update_start("2023-01-01T01:30:00")
+        assert updated.data["start"] == "2023-01-01T01:30:00"
+
+    def test_update_interval_boundary_end(self):
+        interval = Interval.create(Series(
+            {"start": "2023-01-01T01:00:00", "end": "2023-01-01T02:00:00"}
+        ), "start", "end")
+
+        updated = interval.update_end("2023-01-01T02:30:00")
+        assert updated.data["end"] == "2023-01-01T02:30:00"
+
+    def test_update_interval_boundary_return_new_copy(self):
+        interval = Interval.create(Series(
+            {"start": "2023-01-01T01:00:00", "end": "2023-01-01T02:00:00"}
+        ), "start", "end")
+
+        updated = interval.update_start("2023-01-01T01:30:00")
+        assert id(interval) != id(updated)
+        assert interval.data["start"] == "2023-01-01T01:00:00"
+
+    def test_merge_metrics_with_list_metric_merge_true(self):
+        interval = Interval.create(Series({"start": "01:00", "end": "02:00", "value": 10}), "start", "end",
+                                   metric_fields=["value"])
+        other = Interval.create(Series({"start": "01:00", "end": "02:00", "value": 20}), "start", "end",
+                                metric_fields=["value"])
+        expected = Series({"start": "01:00", "end": "02:00", "value": 20})
+
+        merged = interval.merge_metrics(other)
+        assert merged.equals(expected)
+
+    def test_merge_metrics_with_string_metric_column(self):
+        interval = Interval.create(Series({"start": "01:00", "end": "02:00", "value": 10}), "start", "end",
+                                   metric_fields=["value"])
+        other = Interval.create(Series({"start": "01:00", "end": "02:00", "value": 20}), "start", "end",
+                                metric_fields=["value"])
+        expected = Series({"start": "01:00", "end": "02:00", "value": 20})
+
+        merged = interval.merge_metrics(other)
+        assert merged.equals(expected)
+
+    def test_merge_metrics_with_string_metric_columns(self):
+        interval = Interval.create(Series({"start": "01:00", "end": "02:00", "value1": 10, "value2": 20}), "start",
+                                   "end",
+                                   metric_fields=["value1", "value2"])
+        other = Interval.create(Series(
+            {"start": "01:00", "end": "02:00", "value1": 20, "value2": 30}
+        ), "start", "end", metric_fields=["value1", "value2"])
+        expected = Series(
+            {"start": "01:00", "end": "02:00", "value1": 20, "value2": 30}
+        )
+
+        merged = interval.merge_metrics(other)
+        assert merged.equals(expected)
+
+    def test_merge_metrics_return_new_copy(self):
+        interval = Interval.create(
+            Series({"start": "01:00", "end": "02:00", "value": 10}),
+            "start",
+            "end", [], ["value"])
+        other = Interval.create(
+            Series({"start": "01:00", "end": "02:00", "value": 20}),
+            "start",
+            "end", [], ["value"])
+
+        merged = interval.merge_metrics(other)
+        assert id(interval) != id(merged)
+
+    def test_merge_metrics_handle_nan_in_child(self):
+        interval = Interval.create(
+            Series({"start": "01:00", "end": "02:00", "value": 10}),
+            "start",
+            "end",
+            [], ["value"]
+        )
+        other = Interval.create(
+            Series({"start": "01:00", "end": "02:00", "value": float("nan")}),
+            "start",
+            "end",
+            [], ["value"]
+        )
+
+        merged = interval.merge_metrics(other)
+        assert merged["value"] == 10
